Rewrote map system, still funky
