+++ core/src/main/java/org/lpc/MainGame.java
@@ -3,6 +3,7 @@ package org.lpc;
 import com.badlogic.gdx.Application;
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.graphics.OrthographicCamera;
 import lombok.Getter;
 import lombok.extern.log4j.Log4j2;
 import org.apache.logging.log4j.LogManager;
@@ -35,14 +36,14 @@ public class MainGame extends Game {
     public void create() {
         LOGGER.info("Initializing game");
 
-        this.startScreen = new StartScreen(this);
-        setScreen(startScreen);
-
         this.gameStateManager = new GameStateManager(this);
-        this.gameScreen = new GameScreen(this);
-        this.menuScreen = new MenuScreen(this);
         this.inputHandler = new InputHandler(this);
         this.accumulator = 0f;
+
+        this.gameScreen = new GameScreen(this);
+        this.menuScreen = new MenuScreen(this);
+        this.startScreen = new StartScreen(this);
+        setScreen(startScreen);
 
         LOGGER.info("Game created and set to StartScreen");
     }
