+++ core/src/main/java/org/lpc/map/BaseMap.java
@@ -5,6 +5,7 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Vector2;
 import lombok.Getter;
 import lombok.Setter;
+import org.lpc.MainGame;
 import org.lpc.utility.Position;
 
 @Getter
@@ -12,6 +13,11 @@ import org.lpc.utility.Position;
 public abstract class BaseMap {
     protected int width;
     protected int height;
+
+    protected final MapGenerator mapGenerator;
+    protected final MapRenderer renderer;
+
+    private final MainGame game;
 
     // Each map type implements their own tile type
     @Getter
@@ -22,6 +28,14 @@ public abstract class BaseMap {
         public Vector2 getVector2() {
             return new Vector2(position.getGridX(), position.getGridY());
         }
+    }
+
+    public BaseMap(int width, int height, MainGame game) {
+        this.width = width;
+        this.height = height;
+        this.mapGenerator = new MapGenerator();
+        this.renderer = new MapRenderer(game);
+        this.game = game;
     }
 
     public abstract void render(ShapeRenderer shapeRenderer, SpriteBatch batch);
